add delivery price to Order schema

+++ src/main/java/com/orders/routes/stripe/dto/Restaurant.java
@@ -21,7 +21,7 @@ public class Restaurant {
 
     private String country;
 
-    private Integer deliveryPrice;
+    private Long deliveryPrice;
 
     private Integer estimatedDeliveryTime;
